MAINT: Move error computation and saving inside the models
